admin: move 9 forms to Views/
Per #4331:
- Components::AdminSessionForm
    -> Views::Controllers::Admin::Session::Form
- Components::BannerForm
    -> Views::Controllers::Admin::Banners::Form
- Components::DonationForm
    -> Views::Controllers::Admin::Donations::Form
- Components::DonationsReviewForm
    -> Views::Controllers::Admin::Donations::ReviewForm
- Components::IpsManager
    -> Views::Controllers::Admin::BlockedIps::Manager
- Components::MergeRequestForm
    -> Views::Controllers::Admin::Emails::MergeRequests::Form
- Components::NameChangeRequestForm
    -> Views::Controllers::Admin::Emails::NameChangeRequests::Form
- Components::UserBonusesForm
    -> Views::Controllers::Admin::Users::BonusesForm
- Components::WebmasterQuestionForm
    -> Views::Controllers::Admin::Emails::WebmasterQuestions::Form

Admin/table (10th item from the audit) doesn't actually exist; the
audit was stale.

10 ERB callers updated + 3 admin/emails controllers (which use
form_class: in ModalTurboForm dispatch). 36 view tests + 44
controller tests + 10 integration tests pass.

Tightened several forbidden body-wide assertions in the moved test
files to scoped selectors (label[for='...'], .textile, etc.) per
.claude/rules/testing.md. The .textile scope is needed for the .tp
(textile-paragraph) helper's output —  wraps the rendered
text in <div class='textile'>...</div>, so the outer p tag emitted
by  doesn't directly contain the text.
