--- app/components/admin_session_form.rb
@@ -1,27 +1,0 @@
-# frozen_string_literal: true
-
-# Form for admins to switch to another user's account.
-# Allows admins to impersonate users for debugging/support.
-class Components::AdminSessionForm < Components::ApplicationForm
-  def initialize(model, **)
-    super(model, id: "admin_switch_users_form", **)
-  end
-
-  def view_template
-    super do
-      autocompleter_field(
-        :user,
-        type: :user,
-        label: "#{:LOGIN_NAME.l}:",
-        value: model.user,
-        size: 42,
-        autofocus: true
-      )
-      submit(:SUBMIT.l, center: true)
-    end
-  end
-
-  def form_action
-    admin_mode_path
-  end
-end
