+++ app/controllers/admin/emails/webmaster_questions_controller.rb
@@ -19,7 +19,7 @@ module Admin
                      model: FormObject::EmailRequest.new(
                        reply_to: @email, message: @message
                      ),
-                     form_class: Components::WebmasterQuestionForm,
+                     form_class: Views::Controllers::Admin::Emails::WebmasterQuestions::Form,
                      form_locals: { email_error: false }
                    ), layout: false)
           end
