+++ app/views/controllers/admin/banners/index.html.erb
@@ -1,3 +1,3 @@
 <h1><%= :change_banner_title.t %></h1>
 
-<%= render(Components::BannerForm.new(@banner)) %>
+<%= render(Views::Controllers::Admin::Banners::Form.new(@banner)) %>
