+++ app/views/controllers/admin/blocked_ips/edit.html.erb
@@ -11,7 +11,7 @@ container_class(:full)
     end %>
 
     <%# ----- Okay IPs ----- %>
-    <%= render(Components::IpsManager.new(
+    <%= render(Views::Controllers::Admin::BlockedIps::Manager.new(
           FormObject::OkayIps.new,
           type: :okay,
           ips: @okay_ips,
@@ -24,7 +24,7 @@ container_class(:full)
         )) %>
 
     <%# ----- Blocked IPs ----- %>
-    <%= render(Components::IpsManager.new(
+    <%= render(Views::Controllers::Admin::BlockedIps::Manager.new(
           FormObject::BlockedIps.new,
           type: :blocked,
           ips: @blocked_ips,
