+++ app/views/controllers/admin/donations/form.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+module Views::Controllers::Admin::Donations
+  # Form for creating donations (admin only). Rendered by the
+  # admin/donations controller's `new.html.erb`. Allows admins to
+  # manually enter donation information.
+  class Form < ::Components::ApplicationForm
+    def view_template
+      super do
+        text_field(:amount, size: 7, label: "#{:confirm_amount.t}:",
+                            inline: true)
+        text_field(:who, size: 50, label: "#{:WHO.t}:", inline: true)
+        checkbox_field(:anonymous, label: :donate_anonymous.t)
+        text_field(:email, size: 50, label: "#{:EMAIL.t}:", inline: true)
+        submit(:create_donation_add.l, center: true)
+      end
+    end
+
+    def form_action
+      admin_donations_path
+    end
+  end
+end
