+++ app/views/controllers/admin/donations/new.html.erb
@@ -3,4 +3,4 @@ add_page_title(:create_donation_title.l)
 add_context_nav(admin_donations_form_new_tabs)
 %>
 
-<%= render(Components::DonationForm.new(@donation)) %>
+<%= render(Views::Controllers::Admin::Donations::Form.new(@donation)) %>
