+++ app/views/controllers/admin/donations/review_form.rb
@@ -1,0 +1,74 @@
+# frozen_string_literal: true
+
+module Views::Controllers::Admin::Donations
+  # Admin form for reviewing donations. Rendered by the
+  # admin/donations controller's `edit.html.erb`. Sibling to
+  # `Donations::Form` (the new-donation form); the descriptive
+  # suffix distinguishes the two flows.
+  # Renders a table of donations with checkboxes for marking each
+  # donation as reviewed.
+  class ReviewForm < ::Components::ApplicationForm
+    def initialize(form, donations:, **)
+      @donations = donations
+      super(form, id: "admin_review_donations_form",
+                  style: "display: contents", **)
+    end
+
+    def view_template
+      super do
+        submit(:review_donations_update.l, center: true)
+        render_donations_table
+        submit(:review_donations_update.l, center: true)
+      end
+    end
+
+    def form_action
+      admin_donations_path
+    end
+
+    private
+
+    def render_donations_table
+      Table(
+        @donations,
+        class: "table-striped table-review-donations my-3"
+      ) do |t|
+        define_columns(t)
+      end
+    end
+
+    def define_columns(tbl)
+      tbl.column(:review_reviewed.t) do |donation|
+        render_checkbox(donation)
+      end
+      define_donor_columns(tbl)
+      define_detail_columns(tbl)
+    end
+
+    def define_donor_columns(tbl)
+      tbl.column(:review_id.t) { |d| d.id.to_s }
+      tbl.column(:review_who.t, class: "text-left") do |d|
+        d.who.to_s.truncate(30)
+      end
+      tbl.column(:review_anon.t) { |d| d.anonymous.to_s }
+    end
+
+    def define_detail_columns(tbl)
+      tbl.column(:review_amount.t) { |d| d.amount.to_s }
+      tbl.column(:review_email.t, class: "text-left") do |d|
+        d.email.to_s.truncate(50)
+      end
+      tbl.column(:review_date.t) do |d|
+        d.created_at.strftime("%Y-%m-%d")
+      end
+    end
+
+    def render_checkbox(donation)
+      checkbox_field(
+        donation.id.to_s,
+        checked: donation.reviewed,
+        label: false
+      )
+    end
+  end
+end
