+++ app/views/controllers/admin/emails/webmaster_questions/new.erb
@@ -2,7 +2,7 @@
 add_page_title(:ask_webmaster_title.t)
 %>
 
-<%= render(Components::WebmasterQuestionForm.new(
+<%= render(Views::Controllers::Admin::Emails::WebmasterQuestions::Form.new(
       reply_to: @email,
       message: @message,
       email_error: @email_error,
