+++ app/views/controllers/admin/session/edit.html.erb
@@ -5,4 +5,4 @@
 add_page_title(:app_switch_users.l)
 %>
 
-<%= render(Components::AdminSessionForm.new(@form)) %>
+<%= render(Views::Controllers::Admin::Session::Form.new(@form)) %>
