+++ app/views/controllers/admin/session/form.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module Views::Controllers::Admin::Session
+  # Form for admins to switch to another user's account. Rendered by
+  # the admin/session controller's `edit.html.erb`. Allows admins to
+  # impersonate users for debugging/support.
+  class Form < ::Components::ApplicationForm
+    def initialize(model, **)
+      super(model, id: "admin_switch_users_form", **)
+    end
+
+    def view_template
+      super do
+        autocompleter_field(
+          :user,
+          type: :user,
+          label: "#{:LOGIN_NAME.l}:",
+          value: model.user,
+          size: 42,
+          autofocus: true
+        )
+        submit(:SUBMIT.l, center: true)
+      end
+    end
+
+    def form_action
+      admin_mode_path
+    end
+  end
+end
