+++ app/views/controllers/admin/users/bonuses_form.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+module Views::Controllers::Admin::Users
+  # Form for editing user contribution bonuses (admin only).
+  # Rendered by the admin/users controller's `edit.html.erb`. Allows
+  # admins to manually adjust user bonus points. Descriptive
+  # `BonusesForm` name kept since the form is specifically for
+  # bonuses, not generic user edits.
+  class BonusesForm < ::Components::ApplicationForm
+    def view_template
+      super do
+        div(class: "help-note mr-3") { :change_user_bonuses_help.tp }
+        textarea_field(:val, value: model.formatted_bonuses, rows: 5,
+                             class: "mt-3")
+        submit(:SAVE_EDITS.l, center: true)
+      end
+    end
+
+    def form_action
+      admin_user_path(id: model.user_id)
+    rescue NoMethodError
+      # Fallback for tests where admin routes may not be available
+      "/admin/users/#{model.user_id}"
+    end
+  end
+end
